+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ## [v2.1.4] - 2020-03-26
-### Fixed
+### Changed
 - Specify height and hash for rollback explicitly
+### Fixed
+- Make sure that we can't set leaving height smaller than joining height
 
 ## [v2.1.3] - 2020-03-26
 ### Added